authorize user and goal

+++ backend/controllers/goalController.js
@@ -1,12 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
 const { default: mongoose } = require("mongoose");
-
+const User = require("../models/userModel");
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -20,6 +20,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -28,30 +29,58 @@ const setGoal = asyncHandler(async (req, res) => {
 // @route   PUT /api/goals/id
 // @access  Private
 const updateGoal = asyncHandler(async (req, res) => {
-  const goal = Goal.findById(req.params.id);
+  const goal = await Goal.findById(req.params.id);
+
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
   }
-  const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+
+  // Check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
-  res.status(200).json(updateGoal);
+
+  res.status(200).json(updatedGoal);
 });
 
 // @desc    Get goals
 // @route   DELETE /api/goals/id
 // @access  Private
 const deleteGoal = asyncHandler(async (req, res) => {
-  const goal = Goal.findById(req.params.id);
+  const goal = await Goal.findById(req.params.id);
 
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
   }
 
-  const deletedGoal = await Goal.findByIdAndRemove(req.params.id);
-  res.status(200).json({ message: `Goal ${req.params.id} Deleted` });
+  // Check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  await Goal.findByIdAndRemove(req.params.id);
+
+  res.status(200).json({ id: req.params.id });
 });
 
 module.exports = {
